Added executions

+++ src/com/exchange/algo/MatchingResult.java
@@ -1,25 +1,32 @@
 package com.exchange.algo;
 
+/**
+ * Stores result of running Matching Algorithm on an order book
+ *
+ */
 public class MatchingResult {
-	
+
 	/** Whether matching was successful or not */
 	private final boolean matched;
 	/** Price at which match was found */
-	private final double matchingPrice;
+	private final long matchingPrice;
 	/** Volume matched */
 	private final long matchingVolume;
-	
-	public MatchingResult(boolean result, double matchingPrice, long matchingVolume) {
+
+	public MatchingResult(boolean result, long matchingPrice, long matchingVolume) {
 		this.matched = result;
 		this.matchingPrice = matchingPrice;
 		this.matchingVolume = matchingVolume;
 	}
+
 	public boolean matched() {
 		return matched;
 	}
-	public double getMatchingPrice() {
+
+	public long getMatchingPrice() {
 		return matchingPrice;
 	}
+
 	public long getMatchingVolume() {
 		return matchingVolume;
 	}
